aider: Added styles for a delete scene button on each node in the `SceneGraph.scss` file.
# Aider chat conversation:

USER: apps/bot-builder/src/panels/scenes/SceneGraph.scss

ASSISTANT: It seems you've provided the contents of the `SceneGraph.scss` file. If you need any changes to be made to this file, please specify what changes you would like to implement, and I will provide you with the necessary diff.

USER: Can you add adelete scene button for each node? it should be similar to the way the select startSceneId button works, but with a red color and should always display on the scene node hover, if you click it it should remove the scene

ASSISTANT: To add a delete scene button for each node in the `SceneGraph.scss` file, we will need to add new styles for the delete button. Here's the diff for the required changes:

```diff
--- apps/bot-builder/src/panels/scenes/SceneGraph.scss
+++ apps/bot-builder/src/panels/scenes/SceneGraph.scss
@@ ... @@
 .SceneNode__start-icon.selected {
   color: #9747ff;
   opacity: 1;
 }

+// Styles for the delete scene button
+.SceneNode__delete-icon {
+  position: absolute;
+  top: -20px;
+  left: 0px;
+  color: red;
+  cursor: pointer;
+  transition: all 0.2s ease-out;
+  opacity: 0;
+}
+
+.SceneNode__delete-icon:hover {
+  color: darkred;
+  opacity: 1;
+}
+
+.SceneNode:hover .SceneNode__delete-icon {
+  opacity: 0.8;
+}
+
 // Add the following styles
 .react-flow__edge-path {
   cursor: pointer;
 }
```

This diff adds styles for a delete button that will be positioned similarly to the start scene button but on the left side. The button will be red and will become more opaque and darker when hovered over. It will also always display when the scene node is hovered.