aider: Added an "Edit" button to each scene node to trigger opening the modal with the corresponding scene ID.
# Aider chat conversation:

USER: Now, add an EDIT button to each scene node. So in the SCENE NODES you should have on the bottom left area. In the same way that we have the SELECT START SCENE button on each node, we should have an EDIT button, but it should be on the bottom right area. And when you click on that, it should trigger the opening of the model. And you should also set the SCENE EDIT ID to the SCENE I CLICKED.

ASSISTANT: To add an "Edit" button to each scene node and trigger the opening of the modal when clicked, we need to make changes to the `SceneGraph.tsx` file. We will also need to dispatch an action to set the `editId` for the scene in the modal state.

Here are the diffs for the required changes:

```diff
--- apps/bot-builder/src/panels/scenes/SceneGraph.tsx
+++ apps/bot-builder/src/panels/scenes/SceneGraph.tsx
@@ ... @@
import { RiDragMove2Line, RiPlayCircleLine, RiEdit2Line } from "react-icons/ri";
+import { openModal } from "../../state/slices/inputSlice";

@@ ... @@
const SceneNode = ({
  id,
  data,
}: Node<{
  title: string;
  background: string;
  characters: string[];
}>) => {
  const connectionNodeId = useStore((state) => state.connectionNodeId);
  const dispatch = useAppDispatch();
  const startScene = useAppSelector((state) => state.novel.startSceneId);
+ const openSceneEditModal = () => {
+   dispatch(openModal({ modalType: "scene", editId: id }));
+ };
  const setStartSceneId = (id: string) => dispatch(setStartScene(id));

  const isConnecting = !!connectionNodeId;
  const isStartScene = id === startScene;

  return (
    <>
      <div
        className="SceneNode"
        style={{ backgroundImage: `url(${data.background})` }}
      >
        {!isConnecting && (
          <Handle
            className="customHandleSource"
            position={Position.Right}
            type="source"
          />
        )}
        <Handle
          className="customHandleTarget"
          position={Position.Left}
          type="target"
          isConnectableStart={false}
        />
        <div className="SceneNode__title">
          {data.title} <RiDragMove2Line />
        </div>
        <div className="SceneNode__start-icon-container">
          <RiPlayCircleLine
            className={`SceneNode__start-icon ${
              isStartScene ? "selected" : ""
            }`}
            onClick={() => setStartSceneId(id)}
          />
        </div>
+       <div className="SceneNode__edit-icon-container">
+         <RiEdit2Line
+           className="SceneNode__edit-icon"
+           onClick={openSceneEditModal}
+         />
+       </div>
        <div className="SceneNode__characters">
          {data.characters.map((charImg, index) => (
            <img
              key={index}
              src={charImg}
              alt={`Character ${index}`}
              className="SceneNode__character"
            />
          ))}
        </div>
      </div>
    </>
  );
};
```

This diff adds an "Edit" button to the bottom right area of each scene node and dispatches an action to open the `SceneEditModal` with the `editId` set to the ID of the clicked scene.