fixed interaction start not resuming after cutscenes for the first scene
If the first scene had a cutscene, the interaction would not be continued at the end of it, as it only does so if the boolean in cutscene options buffer is set to true. It now sets to true if the user changes scenes or the ai query is stopped.