Fixes 13/10 cutscene inner thoughts (#198)
* Cutscene Displayer: remade skip cutscene

After checking again how custcene part are displayed, I understood that there is no necessity for the use of timeout. I replaced that logic with a while loop, that ends the cutscene on edge cases. Just in case, there is an iteration limit for it.

* TTSPlayer: modified conditional for cached audio playing

When the Inner Thoughts Modal is closed, the cached audio of the TTS Player is cleaned, so there is nothing to play the next time it gets opened. This is fixed by checking for a valid source in the audio element.

* Cutscene Displayer: remade skip cutscene

After checking again how custcene part are displayed, I understood that there is no necessity for the use of timeout. I replaced that logic with a while loop, that ends the cutscene on edge cases. Just in case, there is an iteration limit for it.

* TTSPlayer: modified conditional for cached audio playing

When the Inner Thoughts Modal is closed, the cached audio of the TTS Player is cleaned, so there is nothing to play the next time it gets opened. This is fixed by checking for a valid source in the audio element.

* IndicatorsDisplay: z-index up by one

The inner thoughts trigger' z-index must be 2 so it doesn't get blocked by characters, so I modified indicators display's z-index. It doesn't overlay other components.

* Inner Thoughts: fallback if empty after edit

Replaces the empty inner thought with "..."

* Inner Thoughts: removed fallback

If the inner thought is left empty after edit, a new thought will be generated next time the modal opens.